+++ app/Controllers/UserManagementTokenController.php
@@ -5,6 +5,8 @@ namespace WilokeJWT\Controllers;
 
 
 use WilokeJWT\Helpers\Option;
+use WilokeJWT\Helpers\Session;
+use WilokeJWT\Illuminate\Message\MessageFactory;
 
 class UserManagementTokenController
 {
@@ -13,6 +15,9 @@ class UserManagementTokenController
         add_action('show_user_profile', [$this, 'handleRenewAccessToken']);
         add_action('edit_user_profile', [$this, 'handleRenewAccessToken']);
         add_action('admin_enqueue_scripts', [$this, 'enqueueScripts']);
+        add_action('wp_ajax_wiloke-enable-seen-access-token-user', [$this, 'handleAjaxSeenAcToken']);
+        add_action('wp_ajax_wiloke-enable-seen-refresh-token-user', [$this, 'handleAjaxSeenRfToken']);
+        add_action('wp_ajax_wiloke-renew-token-user', [$this, 'handleAjaxRenewToken']);
     }
 
     public function handleRenewAccessToken()
@@ -63,10 +68,10 @@ class UserManagementTokenController
                         <td class="revoke column-revoke" data-colname="action" style="text-align: center">
                                     <span>
                                         <input type="button"
+                                               class="button button-primary"
                                                value="<?php echo esc_html_e('Renew Token', 'wiloke-jwt'); ?>"
                                                id="wilokeRenewAcToken"
-                                               data-userID="<?php echo esc_attr($userId); ?>"
-                                               style="font-size:20px">
+                                               data-userID="<?php echo esc_attr($userId); ?>">
                                     </span>
                         </td>
                     </tr>
@@ -92,15 +97,88 @@ class UserManagementTokenController
         <?php
     }
 
-    public function enqueueScripts()
+    public function enqueueScripts($hook)
     {
-        wp_enqueue_script('ManagementToken',
-            WILOKE_JWT_URL . 'dist/script.js',
-            ['jquery'],
-            WILOKE_JWT_VERSION,
-            true);
-        wp_localize_script('jquery', 'WILOKE_JWT', [
-            'ajaxurl' => admin_url('admin-ajax.php')
-        ]);
+        if (in_array($hook, ['profile.php', 'user-edit.php'])) {
+            wp_enqueue_script('ManagementToken',
+                WILOKE_JWT_URL . 'dist/script.js',
+                ['jquery'],
+                WILOKE_JWT_VERSION,
+                true);
+            wp_localize_script('jquery', 'WILOKE_JWT', [
+                'ajaxurl' => admin_url('admin-ajax.php')
+            ]);
+        }
+    }
+
+    public function handleAjaxSeenAcToken()
+    {
+        if (current_user_can('administrator') || (get_current_user_id() == ($_GET['user_id'] ?? $_POST['userID']))) {
+            if ($_POST['password'] ?? '') {
+                $oUser = get_userdata(get_current_user_id());
+                if (wp_check_password($_POST['password'], $oUser->data->user_pass, $oUser->ID)) {
+                    Session::sessionStart();
+                    setcookie('enableAcTokenUser' . $_POST['userID'], true, time() + 36000);
+                    return MessageFactory::factory('ajax')->success('The token is enable success', []);
+                } else {
+                    return MessageFactory::factory('ajax')->error('The password incorrect', 401);
+                }
+            } else {
+                return MessageFactory::factory('ajax')->error('Please enter a password', 401);
+            }
+        }else{
+            return MessageFactory::factory('ajax')
+                ->error('The account is not administrator or the current visitor is not a logged in user', 401);
+        }
+        die();
+    }
+
+    public function handleAjaxSeenRfToken()
+    {
+        if (current_user_can('administrator') || (get_current_user_id() == ($_GET['user_id'] ?? $_POST['userID']))) {
+            if ($_POST['password'] ?? '') {
+                $oUser = get_userdata(get_current_user_id());
+                if (wp_check_password($_POST['password'], $oUser->data->user_pass, $oUser->ID)) {
+                    Session::sessionStart();
+                    setcookie('enableRfTokenUser' . $_POST['userID'], true, time() + 36000);
+                    return MessageFactory::factory('ajax')->success('The token is enable success');
+                } else {
+                    return MessageFactory::factory('ajax')->error('The password incorrect', 401);
+                }
+            } else {
+                return MessageFactory::factory('ajax')->error('Please enter a password', 401);
+            }
+        }else{
+            return MessageFactory::factory('ajax')
+                ->error('The account is not administrator or the current visitor is not a logged in user', 401);
+        }
+        die();
+    }
+
+    public function handleAjaxRenewToken()
+    {
+        if ($_POST['password'] ?? '') {
+            $oUser = get_userdata(get_current_user_id());
+            if (wp_check_password($_POST['password'], $oUser->data->user_pass, $oUser->ID)) {
+                if ((get_current_user_id() == $_POST['userId']) || current_user_can('administrator')) {
+                    $aResponse = apply_filters('wiloke/filter/create-access-token-and-refresh-token',
+                        get_userdata($_POST['userId']));
+                    if ($aResponse['code'] == 200) {
+                        Option::saveUserToken($aResponse['accessToken'], $_POST['userId']);
+                        Option::saveUserRefreshToken($aResponse['refreshToken'], $_POST['userId']);
+                        return MessageFactory::factory('ajax')->success('The token is renew success');
+                    }
+                } else {
+                    return MessageFactory::factory('ajax')
+                        ->error('The account is not administrator or the current visitor is not a logged in user', 401);
+                }
+
+            } else {
+                return MessageFactory::factory('ajax')->error('The password incorrect', 401);
+            }
+        } else {
+            return MessageFactory::factory('ajax')->error('Please enter a password', 401);
+        }
+        die();
     }
 }
